deps/v8/test/inspector/debugger/wasm-externref-global-expected.txt

Test wasm scope information with externref globals

Running test: test
Waiting for wasm script to be parsed.
Setting breakpoint in wasm.
Running main.
Paused in debugger.
   globals: {"name": $m.global, "type": externref, "value": hello, world}
